requirements.txt
# Slack
slack-bolt==1.18.0
slack-sdk==3.21.3

# LLM
langchain-groq==0.1.0
langchain-core==0.1.18

# Herramientas
requests==2.31.0

# Configuración
python-dotenv==1.0.0

# Utilidades
typing-extensions==4.8.0
